film_bot_project/requirements.txt
python-telegram-bot==20.8
pymysql
pymongo
tabulate
python-dotenv